+++ client/src/ui/containers/BattleContainer.tsx
@@ -1,0 +1,3 @@
+export const BattleContainer = ({ children }: { children: React.ReactNode }) => {
+  return <div className="w-screen h-screen z-[200] bg-transparent top-0 left-0">{children}</div>;
+};
