+++ client/src/ui/modules/military/BattleView.tsx
@@ -1,0 +1,3 @@
+export const BattleView = () => {
+  return <div className="w-screen"></div>;
+};
