+++ contracts/manifests/dev/manifest.toml
@@ -1941,6 +1941,28 @@ key = false
 
 [[models]]
 kind = "DojoModel"
+class_hash = "0x70b0d7a75f4f826e2d8556abe0c1c0891539611d98658421f9eb2754f263be8"
+original_class_hash = "0x70b0d7a75f4f826e2d8556abe0c1c0891539611d98658421f9eb2754f263be8"
+abi = "abis/deployments/models/eternum_models_tick_tick_move.json"
+name = "eternum::models::tick::tick_move"
+
+[[models.members]]
+name = "entity_id"
+type = "u128"
+key = true
+
+[[models.members]]
+name = "tick"
+type = "u64"
+key = false
+
+[[models.members]]
+name = "count"
+type = "u8"
+key = false
+
+[[models]]
+kind = "DojoModel"
 class_hash = "0x4c92f4209d4c239c7cf0c3d7c645132b2dafd06db44d69f74be5a12bb8a50fb"
 original_class_hash = "0x4c92f4209d4c239c7cf0c3d7c645132b2dafd06db44d69f74be5a12bb8a50fb"
 abi = "abis/deployments/models/eternum_models_trade_status.json"
